refactor: move utils/storage to services/kv-storage
